feat: remove permission module and add find permission methods in role module
