--- user/src/common/enums/permission.events.ts
@@ -1,7 +1,0 @@
-export enum PermissionPatterns {
-  CreatePermission = 'create_permission',
-  GetOnePermission = 'get_one_permission',
-  GetPermissions = 'get_permissions',
-  RemovePermission = 'remove_permission',
-  UpdatePermission = 'update_permission',
-}
