--- user/src/common/enums/permission.messages.ts
@@ -1,7 +1,0 @@
-export enum PermissionMessages {
-  NotFoundPermission = 'Permission not found.',
-  CreatedPermissionSuccess = 'Permission created successfully.',
-  AlreadyExistsPermission = 'Permission already exists with this method and endpoint.',
-  UpdatedPermissionSuccess = 'Permission updated successfully.',
-  RemovedPermissionSuccess = 'Permission removed successfully.',
-}
