+++ user/src/common/interfaces/role.interface.ts
@@ -1,5 +1,4 @@
 import { DefaultRole } from '../enums/shared.enum';
-import { IPermission } from './permission.interface';
 import { IPagination } from './user.interface';
 
 export interface ICreateRole {
@@ -24,6 +23,21 @@ export interface IRolesFilter extends IPagination {
   startDate?: Date;
   endDate?: Date;
   sortBy?: 'name' | 'createdAt' | 'updateAt';
+  sortDirection?: 'asc' | 'desc';
+}
+
+export interface IPermission {
+  method: string;
+  endpoint: string;
+}
+
+export interface IPermissionFilter extends IPagination {
+  method?: string;
+  endpoint?: string;
+  includeRoles?: boolean;
+  startDate?: Date;
+  endDate?: Date;
+  sortBy?: 'endpoint' | 'method' | 'createdAt' | 'updateAt';
   sortDirection?: 'asc' | 'desc';
 }
