+++ user/src/modules/role/role.module.ts
@@ -4,10 +4,9 @@ import { RoleController } from './role.controller';
 import { UserRepository } from '../user/user.repository';
 import { RoleRepository } from './role.repository';
 import { CacheService } from '../cache/cache.service';
-import { PermissionRepository } from '../permission/permission.repository';
 
 @Module({
   controllers: [RoleController],
-  providers: [RoleService, UserRepository, RoleRepository, CacheService, PermissionRepository],
+  providers: [RoleService, UserRepository, RoleRepository, CacheService],
 })
 export class RoleModule {}
