+++ user/src/modules/user/user.module.ts
@@ -10,14 +10,12 @@ import { WalletModule } from '../wallet/wallet.module';
 import { ClientsModule, Transport } from '@nestjs/microservices';
 import { Services } from '../../common/enums/services.enum';
 import { RoleModule } from '../role/role.module';
-import { PermissionModule } from '../permission/permission.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot(envConfig()),
     forwardRef(() => WalletModule),
     forwardRef(() => RoleModule),
-    forwardRef(() => PermissionModule),
     CacheModule,
     PrismaModule,
     ClientsModule.register([
